+++ src/Phan/Analysis/ThrowsTypesAnalyzer.php
@@ -44,7 +44,7 @@ class ThrowsTypesAnalyzer
         /**
          * @param array<int,int|string|Type> $args
          */
-        $maybe_emit_for_method = function(string $issue_type, array $args, Suggestion $suggestion = null) use ($code_base, $method) {
+        $maybe_emit_for_method = function (string $issue_type, array $args, Suggestion $suggestion = null) use ($code_base, $method) {
             if ($method->hasSuppressIssue($issue_type)) {
                 return;
             }