+++ src/Phan/Language/Element/Comment.php
@@ -436,7 +436,7 @@ class Comment
                     // Make sure support for generic types is enabled
                     if (Config::getValue('generic_types_enabled')) {
                         $inherited_type =
-                            self::inheritsFromCommentLine($context, $line);
+                            self::inheritsFromCommentLine($code_base, $context, $line);
                     }
                 } elseif ($type === 'return') {
                     $check_compatible('@return', Comment::FUNCTION_LIKE, $i, $line);
@@ -685,7 +685,8 @@ class Comment
         $return_union_type = UnionType::fromStringInContext(
             $return_union_type_string,
             $context,
-            Type::FROM_PHPDOC
+            Type::FROM_PHPDOC,
+            $code_base
         );
 
         return $return_union_type;
@@ -769,7 +770,8 @@ class Comment
                     UnionType::fromStringInContext(
                         $type,
                         $context,
-                        Type::FROM_PHPDOC
+                        Type::FROM_PHPDOC,
+                        $code_base
                     );
             } else {
                 $union_type = UnionType::empty();
@@ -864,6 +866,9 @@ class Comment
     }
 
     /**
+     * @param CodeBase $code_base
+     * Needed to resolve 'parent', if we ever prefer that shorthand in templates.
+     *
      * @param Context $context
      * The context in which the comment line appears
      *
@@ -874,6 +879,7 @@ class Comment
      * An optional type overriding the extended type of the class
      */
     private static function inheritsFromCommentLine(
+        CodeBase $code_base,
         Context $context,
         string $line
     ) {
@@ -884,7 +890,8 @@ class Comment
             $type = new Some(Type::fromStringInContext(
                 $type_string,
                 $context,
-                Type::FROM_PHPDOC
+                Type::FROM_PHPDOC,
+                $code_base
             ));
 
             return $type;
@@ -922,6 +929,7 @@ class Comment
      * @return ?CommentParameter - if null, the phpdoc magic method was invalid.
      */
     private static function magicParamFromMagicMethodParamString(
+        CodeBase $code_base,
         Context $context,
         string $param_string,
         int $param_index
@@ -941,7 +949,8 @@ class Comment
             $union_type = UnionType::fromStringInContext(
                 $union_type_string,
                 $context,
-                Type::FROM_PHPDOC
+                Type::FROM_PHPDOC,
+                $code_base
             );
             $is_variadic = $param_match[19] === '...';
             $default_str = $param_match[21];
@@ -996,7 +1005,8 @@ class Comment
                     UnionType::fromStringInContext(
                         $return_union_type_string,
                         $context,
-                        Type::FROM_PHPDOC
+                        Type::FROM_PHPDOC,
+                        $code_base
                     );
             } else {
                 // From https://phpdoc.org/docs/latest/references/phpdoc/tags/method.html
@@ -1014,7 +1024,7 @@ class Comment
                 $params_strings = explode(',', $arg_list);
                 $failed = false;
                 foreach ($params_strings as $i => $param_string) {
-                    $param = self::magicParamFromMagicMethodParamString($context, $param_string, $i);
+                    $param = self::magicParamFromMagicMethodParamString($code_base, $context, $param_string, $i);
                     if ($param === null) {
                         Issue::maybeEmit(
                             $code_base,
@@ -1082,7 +1092,8 @@ class Comment
                 UnionType::fromStringInContext(
                     $type,
                     $context,
-                    Type::FROM_PHPDOC
+                    Type::FROM_PHPDOC,
+                    $code_base
                 );
 
             return new CommentParameter(
@@ -1135,7 +1146,8 @@ class Comment
             return new Some(Type::fromStringInContext(
                 $closure_scope_union_type_string,
                 $context,
-                Type::FROM_PHPDOC
+                Type::FROM_PHPDOC,
+                $code_base
             ));
         }
         Issue::maybeEmit(